rename component file using the same style
