+++ client/src/components/Product_page/Product_info/index.js
