+++ client/src/components/Product_page/index.js
@@ -1,7 +1,7 @@
 /* eslint-disable no-unused-vars */
 import React, { Component } from 'react'
 import NavBar from '../Navbar'
-import ProductInfo from '../Product_page/product_info'
+import ProductInfo from '../Product_page/Product_info'
 import AddQuantityCart from '../Product_page/Add_quantity_cart'
 
 class ProductPage extends Component {
